dvbapi: Prevent bogus warning issued by old compilers.
abs() is guaranteed to return positive number so the warning is bogus.

module-dvbapi.c:896:12: warning: comparison of integers of different signs: 'unsigned int' and 'int' [-Wsign-compare]

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@9261 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-dvbapi.c
@@ -893,7 +893,7 @@ int32_t dvbapi_start_emm_filter(int32_t demux_index)
 		cs_debug_mask(D_DVBAPI, "[EMM Filter] %i matching emm filter skipped because they are already active on same emmpid:provid", fcount_added);
 	}
 	if(!fcount) { return 2; }
-	if(fcount == abs(demux[demux_index].emm_filter)) { return 0; }
+	if(fcount == (unsigned int)abs(demux[demux_index].emm_filter)) { return 0; }
 	else { return 1; }
 }
